requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0